new file for 2.0
git-svn-id: https://plugins.svn.wordpress.org/wp-rss-multi-importer/trunk@587049 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -3,6 +3,20 @@
 if( !defined( 'WP_UNINSTALL_PLUGIN' ) )
 exit ();
 // Delete option from options table
-delete_option( 'rss_import_items' );
+if (is_multisite()) {
+    global $wpdb;
+    $blogs = $wpdb->get_results("SELECT blog_id FROM {$wpdb->blogs}", ARRAY_A);
+    if ($blogs) {
+        foreach($blogs as $blog) {
+            switch_to_blog($blog['blog_id']);
+            delete_option('rss_import_items');
+            delete_option('rss_import_categories');
+        }
+        restore_current_blog();
+    }
+} else {
+    delete_option('rss_import_items');
+    delete_option('rss_import_categories');
+}
 //
 ?>
